Add IDF 5 support and deprecate versions 4.1 and 4.2 (#176)
* Update support for ESP 5


* C_STANDARD workarround


* Update docker, nightly and readme


* Generalice C_VERSION for different IDF versions


* Remove C_STANDARD from toolchain


* Fix embeddeRTPS


* Fix example


* Fix esp32c3 build


* Fix macro name


* Deprecate v4.1 and v4.2


* Deprecate galactic


* Revert XRCE client temporal branch


---------

(cherry picked from commit a71fb638eb89d9ccf97b7142b29ec981986a0d44)

# Conflicts:
#	.github/workflows/nightly.yml

+++ examples/int32_publisher_embeddedrtps/main/main.c
@@ -24,7 +24,7 @@ void timer_callback(rcl_timer_t * timer, int64_t last_call_time)
 {
 	RCLC_UNUSED(last_call_time);
 	if (timer != NULL) {
-		printf("Publishing: %d\n", msg.data);
+		printf("Publishing: %d\n", (int) msg.data);
 		RCSOFTCHECK(rcl_publish(&publisher, &msg, NULL));
 		msg.data++;
 	}
