+++ examples/int32_sub_pub/main/main.c
@@ -31,7 +31,7 @@ void timer_callback(rcl_timer_t * timer, int64_t last_call_time)
 	(void) last_call_time;
 	if (timer != NULL) {
 		RCSOFTCHECK(rcl_publish(&publisher, &send_msg, NULL));
-		printf("Sent: %d\n", send_msg.data);
+		printf("Sent: %d\n",  (int)  send_msg.data);
 		send_msg.data++;
 	}
 }
@@ -39,7 +39,7 @@ void timer_callback(rcl_timer_t * timer, int64_t last_call_time)
 void subscription_callback(const void * msgin)
 {
 	const std_msgs__msg__Int32 * msg = (const std_msgs__msg__Int32 *)msgin;
-	printf("Received: %d\n", msg->data);
+	printf("Received: %d\n",  (int)  msg->data);
 }
 
 void micro_ros_task(void * arg)
